+++ development/1.0.2/src/dev/hubpilot/interact/HubPilotInteractPlugin.java
@@ -144,4 +144,3 @@ extends JavaPlugin {
 
     }
 }
-